Move test cases to IPv4_Sockets
