+++ test/unit-test/FreeRTOS_Sockets/FreeRTOS_Sockets_GenericAPI_utest.c
@@ -127,103 +127,6 @@ static BaseType_t xLocalReceiveCallback( Socket_t xSocket,
     return xLocalReceiveCallback_Return;
 }
 
-
-/*
- * @brief Test for FreeRTOS_inet_pton4 function.
- */
-void test_FreeRTOS_inet_pton4( void )
-{
-    char * pucValidString1 = "192.68.1.1";
-    uint32_t ulValidResponse1 = 0x010144C0;
-    char * pucValidString2 = "192.0.1.1";
-    uint32_t ulValidResponse2 = 0x010100C0;
-    char * pucValidString3 = "192.68.1.0";
-    uint32_t ulValidResponse3 = 0x000144C0;
-    char * pucValidString4 = "0.0.0.0";
-    uint32_t ulValidResponse4 = 0x00000000;
-    char * pucValidString5 = "0.68.1.1";
-    uint32_t ulValidResponse5 = 0x01014400;
-
-    char * pucInvalidString1 = "0192.68.1.1";
-    char * pucInvalidString2 = "192.68.00.1";
-    char * pucInvalidString3 = "192.00.1.1";
-    char * pucInvalidString4 = "freertos.com";
-    char * pucInvalidString5 = "192.freertos.com";
-    char * pucInvalidString6 = "\0";
-    char * pucInvalidString7 = "1234.68.1.1";
-    char * pucInvalidString8 = "123.68.0a.1";
-    uint32_t ulInValidResponse = 0x00;
-
-    BaseType_t xResult;
-    uint32_t ulIPAddress;
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucValidString1, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdPASS, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulValidResponse1, ulIPAddress, "Could not convert string 1 correctly." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucValidString2, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdPASS, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulValidResponse2, ulIPAddress, "Could not convert string 2 correctly." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucValidString3, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdPASS, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulValidResponse3, ulIPAddress, "Could not convert string 3 correctly." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucValidString4, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdPASS, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulValidResponse4, ulIPAddress, "Could not convert string 4 correctly." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucValidString5, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdPASS, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulValidResponse5, ulIPAddress, "Could not convert string 5 correctly." );
-
-    /* Invalid test cases. */
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucInvalidString1, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdFAIL, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulInValidResponse, ulIPAddress, "Incorrectly converted string 1." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucInvalidString2, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdFAIL, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulInValidResponse, ulIPAddress, "Incorrectly converted string 2." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucInvalidString3, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdFAIL, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulInValidResponse, ulIPAddress, "Incorrectly converted string 3." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucInvalidString4, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdFAIL, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulInValidResponse, ulIPAddress, "Incorrectly converted string 4." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucInvalidString5, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdFAIL, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulInValidResponse, ulIPAddress, "Incorrectly converted string 5." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucInvalidString6, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdFAIL, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulInValidResponse, ulIPAddress, "Incorrectly converted string 6." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucInvalidString7, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdFAIL, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulInValidResponse, ulIPAddress, "Incorrectly converted string 7." );
-
-    ulIPAddress = 0xABABABAB;
-    xResult = FreeRTOS_inet_pton4( pucInvalidString8, &ulIPAddress );
-    TEST_ASSERT_EQUAL( pdFAIL, xResult );
-    TEST_ASSERT_EQUAL_MESSAGE( ulInValidResponse, ulIPAddress, "Incorrectly converted string 8." );
-}
-
 /*
  * @brief Creation of socket when socket size determination fails as IP task is not ready.
  */
@@ -2365,68 +2268,6 @@ void test_FreeRTOS_inet_pton_HappyPath( void )
 
     TEST_ASSERT_EQUAL( pdPASS, xReturn );
     TEST_ASSERT_EQUAL_UINT32( 0xFFFFFFFF, ulDestination );
-}
-
-/*
- * @brief Incorrect family of address.
- */
-void test_FreeRTOS_inet_ntop_IncorrectAddrFamily( void )
-{
-    char * pucReturn;
-    BaseType_t xAddressFamily = FREERTOS_AF_INET + 1;
-    uint32_t ulSource;
-    char * pcDestination;
-    socklen_t uxSize = 100;
-
-    pucReturn = FreeRTOS_inet_ntop( xAddressFamily, &ulSource, pcDestination, uxSize );
-
-    TEST_ASSERT_EQUAL( NULL, pucReturn );
-}
-
-/*
- * @brief Incorrect size of the string being passed to the function.
- */
-void test_FreeRTOS_inet_ntop_IncorrectSize( void )
-{
-    char * pucReturn;
-    BaseType_t xAddressFamily = FREERTOS_AF_INET;
-    uint32_t ulSource;
-    const socklen_t uxSize = 15;
-    char pcDestination[ uxSize ];
-
-    pucReturn = FreeRTOS_inet_ntop( xAddressFamily, &ulSource, pcDestination, uxSize );
-
-    TEST_ASSERT_EQUAL( NULL, pucReturn );
-}
-
-/*
- * @brief Happy path.
- */
-void test_FreeRTOS_inet_ntop_HappyCase( void )
-{
-    char * pucReturn;
-    BaseType_t xAddressFamily = FREERTOS_AF_INET;
-    uint32_t ulSource;
-    const socklen_t uxSize = 16;
-    char pcDestination[ uxSize ];
-
-    ulSource = 0xFFFFFFFF;
-    pucReturn = FreeRTOS_inet_ntop( xAddressFamily, &ulSource, pcDestination, uxSize );
-
-    TEST_ASSERT_EQUAL( pcDestination, pucReturn );
-    TEST_ASSERT_EQUAL_STRING( "255.255.255.255", pcDestination );
-
-    ulSource = 0;
-    pucReturn = FreeRTOS_inet_ntop( xAddressFamily, &ulSource, pcDestination, uxSize );
-
-    TEST_ASSERT_EQUAL( pcDestination, pucReturn );
-    TEST_ASSERT_EQUAL_STRING( "0.0.0.0", pcDestination );
-
-    ulSource = 0xABCDEF12;
-    pucReturn = FreeRTOS_inet_ntop( xAddressFamily, &ulSource, pcDestination, uxSize );
-
-    TEST_ASSERT_EQUAL( pcDestination, pucReturn );
-    TEST_ASSERT_EQUAL_STRING( "18.239.205.171", pcDestination );
 }
 
 /*
